+++ src/PeakPresence/DiscordRPCPatch.cs
@@ -3,13 +3,11 @@ using DiscordRPC;
 using Photon.Pun;
 using System;
 using WebSocketSharp;
-using AncestralMod;
 
 namespace PeakPresence;
 
 class DiscordRPCPatch
 {
-
 
 	[HarmonyPatch(typeof(RichPresenceService), "SetState")]
 	[HarmonyPostfix]
@@ -25,28 +23,6 @@ class DiscordRPCPatch
 		UpdateDiscordRPC(__instance._presence.State);
 	}
 
-	// private static Character Player => Character.localCharacter;
-	
-	// [HarmonyPatch(typeof(Character), "HandleDeath")]
-	// [HarmonyPrefix]
-	// static void HandleDeathPrefix(Character __instance)
-	// {
-	// 	if (__instance.data.sinceDied == 0f)
-	// 	{
-	// 		RichPresenceService? richPresence = GameHandler.GetService<RichPresenceService>();
-	// 		if (richPresence != null) UpdateDiscordRPC(richPresence.m_currentState);
-	// 	}
-	// }
-	
-	// [HarmonyPatch(typeof(Character), "HandlePassedOut")]
-	// [HarmonyPrefix]
-	// static void HandlePassedOutPrefix(Character __instance)
-	// {
-	// 	if (__instance.data.lastPassedOut == 0f) {
-	// 		RichPresenceService? richPresence = GameHandler.GetService<RichPresenceService>();
-	// 		if (richPresence != null) UpdateDiscordRPC(richPresence.m_currentState);
-	// 	}
-	// }
 	
 	public static void UpdateDiscordRPC(RichPresenceState currentState) {
 		if (Plugin.Client != null)
@@ -58,30 +34,6 @@ class DiscordRPCPatch
 
 			string State = "";
 			Party? Party = null;
-
-			// if (Helper.IsOnIsland())
-			// {
-			// 	if (string.IsNullOrEmpty(State) && ConfigHandler.ShowAliveStatus.Value)
-			// 	{
-			// 		if (Player.data.dead) State = LocalizationManager.Get("status.dead");
-			// 		else if (Player.data.passedOut) State = LocalizationManager.Get("status.passed_out");
-			// 	}
-			// 	if (string.IsNullOrEmpty(State) && ConfigHandler.ShowHeight.Value)
-			// 	{
-			// 		float? height = Player.refs.stats.heightInMeters;
-			// 		if (height != null)
-			// 		{
-			// 			float heightValue = height.Value;
-			// 			string unit = LocalizationManager.Get("progress.height.meters");
-			// 			if (heightValue >= 1000f && ConfigHandler.AbbreviateHeight.Value)
-			// 			{
-			// 				heightValue /= 1000f;
-			// 				unit = LocalizationManager.Get("progress.height.kilometers");
-			// 			}
-			// 			State = string.Format(LocalizationManager.Get(unit), heightValue.ToString("F2"));
-			// 		}
-			// 	}
-			// }
 
 			if (string.IsNullOrEmpty(State)) State = PhotonNetwork.OfflineMode ? LocalizationManager.Get("playing.solo") : PhotonNetwork.InRoom ? LocalizationManager.Get("playing.multiplayer") : "";
 			if (!PhotonNetwork.OfflineMode && PhotonNetwork.InRoom)
@@ -104,25 +56,14 @@ class DiscordRPCPatch
 			float? currentTime = Helper.GetCurrentGameTime();
 			if (currentTime != null)
 				Timestamps = new Timestamps { Start = DateTime.UtcNow.AddSeconds(-currentTime.Value) };
-			Plugin.Client.SetPresence(
-				new RichPresence()
-				{
-					Details = Details,
-					State = State,
-					Party = Party,
-					Assets = Assets,
-					Timestamps = Timestamps,
-					Type = ActivityType.Playing
-				}
-			);
-			// DiscordRPCQueue.SendRefresh(new DiscordRPCQueue.RefreshRequest(
-			// 	Details,
-			// 	State,
-			// 	Party,
-			// 	Assets,
-			// 	Timestamps,
-			// 	ActivityType.Playing
-			// ));
+			DiscordRPCQueue.SendRefresh(new DiscordRPCQueue.RefreshRequest(
+				Details,
+				State,
+				Party,
+				Assets,
+				Timestamps,
+				ActivityType.Playing
+			));
 		}
 	}